refactor(sdk): remove all version references from codebase
