+++ brokle/client.py
@@ -1,10 +1,10 @@
 """
-Main client for Brokle SDK v2.0.
+Main client for Brokle SDK.
 
 This module provides the main client interface with OpenAI-compatible design
 and clean resource management.
 
-v2.0 Changes:
+Key Features:
 - Clean OpenAI-compatible interface
 - Proper resource management with context managers
 - Sync and async client variants
